chore(test): add separate file for getting params
