+++ test/chains/cosmos/localnet.go
@@ -2,12 +2,10 @@ package cosmos
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"strings"
 	"testing"
+	"time"
 
-	"github.com/cosmos/cosmos-sdk/types"
 	"github.com/icon-project/ibc-integration/test/chains"
 	"github.com/icon-project/ibc-integration/test/internal/blockdb"
 	"github.com/strangelove-ventures/interchaintest/v6/chain/cosmos"
@@ -57,6 +55,9 @@ func (c *CosmosLocalnet) DeployContract(ctx context.Context, keyName string) (co
 }
 
 func (c *CosmosLocalnet) QueryContract(ctx context.Context, contractAddress, methodName, params string) (context.Context, error) {
+	// wait for few blocks after executing before querying
+	time.Sleep(2 * time.Second)
+
 	// get query msg
 	queryMsg := c.GetQueryParam(methodName)
 	chains.Response = ""
@@ -88,90 +89,8 @@ func (c *CosmosLocalnet) FindTxs(ctx context.Context, height uint64) ([]blockdb.
 	return nil, nil
 }
 
-func (c *CosmosLocalnet) SetAdminParams(ctx context.Context, keyName string) (context.Context, string, error) {
-	var admin Admin
-	var admins chains.Admins
-	originalJSON := `{"set_admin":{"address":""}}`
-	json.Unmarshal([]byte(originalJSON), &admin)
-	if strings.ToLower(keyName) == "null" {
-		return context.WithValue(ctx, chains.AdminKey("Admins"), chains.Admins{
-			Admin: admins.Admin,
-		}), string(originalJSON), nil
-	} else if strings.ToLower(keyName) == "junk" {
-		admin.SetAdmin.Address = "$%#^!(&^%^)"
-		updatedJSON, _ := json.Marshal(admin)
-		return context.WithValue(ctx, chains.AdminKey("Admins"), chains.Admins{
-			Admin: admins.Admin,
-		}), string(updatedJSON), nil
-	} else {
-		// Check if the given wallet exists if not create a wallet
-		addr, err := c.CosmosChain.GetAddress(ctx, keyName)
-		if err != nil {
-			c.BuildWallets(ctx, keyName)
-			addr, _ = c.CosmosChain.GetAddress(ctx, keyName)
-		}
-		adminAddr, _ := types.Bech32ifyAddressBytes(c.CosmosChain.Config().Bech32Prefix, addr)
-		admin.SetAdmin.Address = adminAddr
-		updatedJSON, _ := json.Marshal(admin)
-		fmt.Println(string(updatedJSON))
-		admins.Admin = map[string]string{
-			keyName: adminAddr,
-		}
-		return context.WithValue(ctx, chains.AdminKey("Admins"), chains.Admins{
-			Admin: admins.Admin,
-		}), string(updatedJSON), nil
-	}
-}
-
-func (c *CosmosLocalnet) GetQueryParam(method string) Query {
-	var queryMsg Query
-	if strings.Contains(method, "admin") {
-		queryMsg = Query{GetAdmin: &GetAdmin{}}
-	} else if strings.Contains(method, "fee") {
-		queryMsg = Query{GetProtocolFee: &GetProtocolFee{}}
-	}
-	return queryMsg
-}
-
-func (c *CosmosLocalnet) GetExecuteParam(ctx context.Context, methodaName, param string) (context.Context, string, error) {
-	if strings.Contains(methodaName, "admin") {
-		return c.SetAdminParams(ctx, param)
-	}
-	return ctx, "", nil
-}
-
 func (c *CosmosLocalnet) BuildWallets(ctx context.Context, keyName string) error {
 	// Build Wallet and fund user
 	_, err := c.GetAndFundTestUser(ctx, keyName, int64(100_000_000), c.CosmosChain)
 	return err
-}
-
-func (c *CosmosLocalnet) GetAndFundTestUser(
-	ctx context.Context,
-	keyNamePrefix string,
-	amount int64,
-	chain ibc.Chain,
-) (string, error) {
-	// Check if the address for the given key is already created
-	_, err := c.CosmosChain.GetAddress(ctx, keyNamePrefix)
-	if err != nil {
-		chainCfg := c.CosmosChain.Config()
-		user, err := chain.BuildWallet(ctx, keyNamePrefix, "")
-		if err != nil {
-			return "", fmt.Errorf("failed to get source user wallet: %w", err)
-		}
-
-		err = chain.SendFunds(ctx, chains.FaucetAccountKeyName, ibc.WalletAmount{
-			Address: user.FormattedAddress(),
-			Amount:  amount,
-			Denom:   chainCfg.Denom,
-		})
-		if err != nil {
-			return "", fmt.Errorf("failed to get funds from faucet: %w", err)
-		}
-		fmt.Printf("Address of %s is : %s \n", user.KeyName(), user.FormattedAddress())
-		return user.KeyName(), nil
-	} else {
-		return keyNamePrefix, err
-	}
 }
